requirements.txt
# Global requierements
configparser
datetime
python-dotenv


# Project specific
fritzconnection
psycopg
pandas
schedule